refactor(wizards): rename foundation folder
